// apps/webcert/src/types/resourceLink.ts
export interface ResourceLink {
  type: ResourceLinkType
  name: string
  title?: string
  description: string
  body?: string
  enabled: boolean
}

export enum ResourceLinkType {
  ACCESS_DRAFT_LIST = 'ACCESS_DRAFT_LIST',
  ACCESS_QUESTION_LIST = 'ACCESS_QUESTION_LIST',
  ACCESS_SEARCH_CREATE_PAGE = 'ACCESS_SEARCH_CREATE_PAGE',
  ACCESS_SIGNED_CERTIFICATES_LIST = 'ACCESS_SIGNED_CERTIFICATES_LIST',
  ANSWER_QUESTION = 'ANSWER_QUESTION',
  CANNOT_COMPLEMENT_CERTIFICATE = 'CANNOT_COMPLEMENT_CERTIFICATE',
  CANNOT_COMPLEMENT_CERTIFICATE_ONLY_MESSAGE = 'CANNOT_COMPLEMENT_CERTIFICATE_ONLY_MESSAGE',
  CHANGE_UNIT = 'CHANGE_UNIT',
  CHOOSE_UNIT = 'CHOOSE_UNIT',
  COMPLEMENT_CERTIFICATE = 'COMPLEMENT_CERTIFICATE',
  COPY_CERTIFICATE = 'COPY_CERTIFICATE',
  COPY_CERTIFICATE_CONTINUE = 'COPY_CERTIFICATE_CONTINUE',
  CREATE_CERTIFICATE = 'CREATE_CERTIFICATE',
  CREATE_CERTIFICATE_FROM_CANDIDATE = 'CREATE_CERTIFICATE_FROM_CANDIDATE',
  CREATE_CERTIFICATE_FROM_CANDIDATE_WITH_MESSAGE = 'CREATE_CERTIFICATE_FROM_CANDIDATE_WITH_MESSAGE',
  CREATE_CERTIFICATE_FROM_TEMPLATE = 'CREATE_CERTIFICATE_FROM_TEMPLATE',
  CREATE_DEATHCERTIFICATE_CONFIRMATION = 'CREATE_DEATHCERTIFICATE_CONFIRMATION',
  CREATE_LUAENA_CONFIRMATION = 'CREATE_LUAENA_CONFIRMATION',
  CREATE_QUESTIONS = 'CREATE_QUESTIONS',
  DISPLAY_PATIENT_ADDRESS_IN_CERTIFICATE = 'DISPLAY_PATIENT_ADDRESS_IN_CERTIFICATE',
  EDIT_CERTIFICATE = 'EDIT_CERTIFICATE',
  FMB = 'FMB',
  FORWARD_CERTIFICATE = 'FORWARD_CERTIFICATE',
  FORWARD_QUESTION = 'FORWARD_QUESTION',
  HANDLE_QUESTION = 'HANDLE_QUESTION',
  LOG_OUT = 'LOG_OUT',
  MISSING_RELATED_CERTIFICATE_CONFIRMATION = 'MISSING_RELATED_CERTIFICATE_CONFIRMATION',
  NAVIGATE_BACK_BUTTON = 'NAVIGATE_BACK_BUTTON',
  PRINT_CERTIFICATE = 'PRINT_CERTIFICATE',
  PRIVATE_PRACTITIONER_PORTAL = 'PRIVATE_PRACTITIONER_PORTAL',
  QUESTIONS = 'QUESTIONS',
  QUESTIONS_ADMINISTRATIVE = 'QUESTIONS_ADMINISTRATIVE',
  QUESTIONS_NOT_AVAILABLE = 'QUESTIONS_NOT_AVAILABLE',
  READ_CERTIFICATE = 'READ_CERTIFICATE',
  READY_FOR_SIGN = 'READY_FOR_SIGN',
  REMOVE_CERTIFICATE = 'REMOVE_CERTIFICATE',
  RENEW_CERTIFICATE = 'RENEW_CERTIFICATE',
  REPLACE_CERTIFICATE = 'REPLACE_CERTIFICATE',
  REPLACE_CERTIFICATE_CONTINUE = 'REPLACE_CERTIFICATE_CONTINUE',
  REVOKE_CERTIFICATE = 'REVOKE_CERTIFICATE',
  SEND_CERTIFICATE = 'SEND_CERTIFICATE',
  SHOW_RELATED_CERTIFICATE = 'SHOW_RELATED_CERTIFICATE',
  SIGN_CERTIFICATE = 'SIGN_CERTIFICATE',
  SIGN_CERTIFICATE_CONFIRMATION = 'SIGN_CERTIFICATE_CONFIRMATION',
  SRS_FULL_VIEW = 'SRS_FULL_VIEW',
  SRS_MINIMIZED_VIEW = 'SRS_MINIMIZED_VIEW',
  SUBSCRIPTION_WARNING = 'SUBSCRIPTION_WARNING',
  WARNING_LUAENA_INTEGRATED = 'WARNING_LUAENA_INTEGRATED',
  WARNING_NORMAL_ORIGIN = 'WARNING_NORMAL_ORIGIN',
}